+++ packages/core/src/providers/pinpoint/utils/getEventBuffer.ts
